+++ app/src/main/java/com/yundian/star/ui/main/adapter/BookStarListAdapter.java
@@ -2,6 +2,7 @@ package com.yundian.star.ui.main.adapter;
 
 import android.content.Context;
 import android.view.View;
+import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.netease.nim.uikit.NimUIKit;
@@ -10,7 +11,13 @@ import com.yundian.star.R;
 import com.yundian.star.base.ListBaseAdapter;
 import com.yundian.star.base.SuperViewHolder;
 import com.yundian.star.been.BookingStarListBean;
+import com.yundian.star.greendao.GreenDaoManager;
+import com.yundian.star.greendao.StarInfo;
 import com.yundian.star.ui.wangyi.session.activity.P2PMessageActivity;
+import com.yundian.star.utils.ImageLoaderUtils;
+
+import java.util.List;
+import static com.yundian.star.R.id.tv_star_name;
 
 /**
  * Created by sll on 2017/5/24.
@@ -29,12 +36,13 @@ public class BookStarListAdapter extends ListBaseAdapter<BookingStarListBean> {
     @Override
     public void onBindItemHolder(SuperViewHolder holder, final int position) {
         final BookingStarListBean item = mDataList.get(position);
-        TextView nameText = holder.getView(R.id.tv_star_name);
+        TextView nameText = holder.getView(tv_star_name);
         TextView bookinged = holder.getView(R.id.tv_booking);
         TextView refuse = holder.getView(R.id.tv_booking_refuse);
         TextView talk = holder.getView(R.id.have_talk);
+        ImageView iv_star_head = holder.getView(R.id.iv_star_head);
         nameText.setText(item.getStarname());
-        if (item.getStatus() == 1) {
+        if (item.getAppoint() == 1) {
             bookinged.setVisibility(View.VISIBLE);
             refuse.setVisibility(View.GONE);
         } else {
@@ -48,6 +56,11 @@ public class BookStarListAdapter extends ListBaseAdapter<BookingStarListBean> {
                 P2PMessageActivity.start(mContext, item.getFaccid(),item.getStarcode(), customization, null);
             }
         });
+        List<StarInfo> starInfos = GreenDaoManager.getInstance().queryLove(item.getStarcode());
+        if (starInfos!=null&&starInfos.size()!=0){
+            StarInfo starInfo = starInfos.get(0);
+            ImageLoaderUtils.display(mContext,iv_star_head,starInfo.getPic_url());
+        }
 
     }
 }
